+++ Pinta.Core/Actions/WindowActions.cs
@@ -31,16 +31,13 @@ public sealed class WindowActions
 	private Gio.Menu doc_section = null!; // NRT - Set in RegisterActions
 	private static readonly string doc_action_id = "active_document";
 	private readonly Gio.SimpleAction active_doc_action;
+	private uint deferred_update_event;
 
 	public Command SaveAll { get; }
 	public Command CloseAll { get; }
 
-	private readonly ChromeManager chrome;
 	private readonly WorkspaceManager workspace;
-	public WindowActions (
-		ChromeManager chrome,
-		ToolManager tools,
-		WorkspaceManager workspace)
+	public WindowActions (WorkspaceManager workspace)
 	{
 		SaveAll = new Command (
 			"SaveAll",
@@ -63,16 +60,8 @@ public sealed class WindowActions
 			GLib.Variant.NewInt32 (-1));
 
 		active_doc_action.OnActivate += (o, e) => {
-
-			var idx = e.Parameter!.GetInt32 ();
-
-			if (idx >= workspace.OpenDocuments.Count)
-				return;
-
-			workspace.SetActiveDocumentInternal (
-				tools,
-				workspace.OpenDocuments[idx]);
-
+			int idx = e.Parameter!.GetInt32 ();
+			workspace.SetActiveDocument (idx);
 			active_doc_action.ChangeState (e.Parameter);
 		};
 
@@ -81,9 +70,10 @@ public sealed class WindowActions
 			e.Document.IsDirtyChanged += (_, _) => RebuildDocumentMenu ();
 			AddDocumentMenuItem (workspace.OpenDocuments.IndexOf (e.Document));
 		};
+
+		workspace.ActiveDocumentChanged += OnActiveDocumentChanged;
 		workspace.DocumentClosed += (_, _) => RebuildDocumentMenu ();
 
-		this.chrome = chrome;
 		this.workspace = workspace;
 	}
 
@@ -102,27 +92,23 @@ public sealed class WindowActions
 			CloseAll]);
 
 		app.AddAction (active_doc_action);
-	}
 
-	public void SetActiveDocument (Document doc)
-	{
-		int idx = workspace.OpenDocuments.IndexOf (doc);
-		active_doc_action.Activate (GLib.Variant.NewInt32 (idx));
+		// Assign accelerators up to Alt-9 for the active documents.
+		for (int i = 0; i < 9; ++i)
+			app.SetAccelsForAction (BuildActionId (i), [$"<Alt>{i + 1}"]);
 	}
 
 	private void AddDocumentMenuItem (int idx)
 	{
 		Document doc = workspace.OpenDocuments[idx];
-		string action_id = $"app.{doc_action_id}({idx})";
-		string label = $"{doc.DisplayName}{(doc.IsDirty ? '*' : string.Empty)}";
-		Gio.MenuItem menu_item = Gio.MenuItem.New (label, action_id);
+		string actionId = BuildActionId (idx);
+		string label = $"{doc.DisplayName}{(doc.IsDirty ? "*" : string.Empty)}";
 
-		doc_section.AppendItem (menu_item);
-
-		// We only assign accelerators up to Alt-9
-		if (idx < 9)
-			chrome.Application.SetAccelsForAction (action_id, [$"<Alt>{idx + 1}"]);
+		Gio.MenuItem menuItem = Gio.MenuItem.New (label, actionId);
+		doc_section.AppendItem (menuItem);
 	}
+
+	private static string BuildActionId (int idx) => $"app.{doc_action_id}({idx})";
 
 	private void RebuildDocumentMenu ()
 	{
@@ -132,5 +118,19 @@ public sealed class WindowActions
 			AddDocumentMenuItem (i);
 
 		workspace.ResetTitle ();
+	}
+
+	private void OnActiveDocumentChanged (object? o, System.EventArgs eventArgs)
+	{
+		// Updating the action's state can be surprisingly expensive when e.g. opening
+		// many documents (bug #1574), so an update is deferred until we return to the event loop.
+		if (deferred_update_event > 0)
+			return;
+
+		deferred_update_event = GLib.Functions.IdleAdd (GLib.Constants.PRIORITY_DEFAULT, () => {
+			active_doc_action.ChangeState (GLib.Variant.NewInt32 (workspace.ActiveDocumentIndex));
+			deferred_update_event = 0;
+			return false;
+		});
 	}
 }