+++ Pinta.Core/Classes/AsyncEffectRenderer.cs
@@ -30,8 +30,6 @@
 
 using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -40,88 +38,39 @@ using Debug = System.Diagnostics.Debug;
 namespace Pinta.Core;
 
 // Only call methods on this class from a single thread (The UI thread).
-internal sealed class AsyncEffectRenderer : IDisposable
+internal static class AsyncEffectRenderer
 {
-	internal readonly record struct CompletionInfo (
-		bool WasCanceled,
-		IReadOnlyList<Exception> Errors);
-
 	internal sealed class Settings
 	{
 		internal int ThreadCount { get; }
 		internal RectangleI RenderBounds { get; }
 		internal bool EffectIsTileable { get; }
-		internal int UpdateMillis { get; }
-		internal ThreadPriority ThreadPriority { get; }
 
 		internal Settings (
 			int threadCount,
 			RectangleI renderBounds,
-			bool effectIsTileable,
-			int updateMilliseconds,
-			ThreadPriority threadPriority)
+			bool effectIsTileable)
 		{
 			if (renderBounds.Width < 0) throw new ArgumentException ("Width cannot be negative", nameof (renderBounds));
 			if (renderBounds.Height < 0) throw new ArgumentException ("Height cannot be negative", nameof (renderBounds));
-			if (updateMilliseconds <= 0) throw new ArgumentOutOfRangeException (nameof (updateMilliseconds), "Strictly positive value expected");
 			if (threadCount < 1) throw new ArgumentOutOfRangeException (nameof (threadCount), "Invalid number of threads");
 			RenderBounds = renderBounds;
 			EffectIsTileable = effectIsTileable;
 			ThreadCount = threadCount;
-			UpdateMillis = updateMilliseconds;
-			ThreadPriority = threadPriority;
 		}
 	}
 
-	TaskCompletionSource<CompletionInfo> completion_source;
-	CancellationTokenSource cancellation_source;
-	ConcurrentQueue<RectangleI> queued_tiles;
-	int tiles_count = 0;
-
-	uint timer_tick_id;
-
-	RectangleI updated_area;
-
-	private readonly Settings settings;
-
-	internal AsyncEffectRenderer (Settings settings)
-	{
-		TaskCompletionSource<CompletionInfo> initialCompletionSource = new ();
-		initialCompletionSource.SetResult (
-			new (
-				WasCanceled: false,
-				Errors: []
-			)
-		);
-
-		completion_source = initialCompletionSource;
-		cancellation_source = new ();
-		queued_tiles = new ConcurrentQueue<RectangleI> ();
-
-		timer_tick_id = 0;
-
-		this.settings = settings;
-	}
-
-	internal bool IsRendering => !completion_source.Task.IsCompleted;
-
-	internal double Progress {
-		get {
-			if (tiles_count == 0) return 0;
-			int dequeued = tiles_count - queued_tiles.Count;
-			return dequeued / (double) tiles_count;
-		}
-	}
-
-	internal void Start (
+	internal static RenderHandle Start (
+		Settings settings,
 		BaseEffect effect,
 		Cairo.ImageSurface source,
 		Cairo.ImageSurface dest)
 	{
-		if (IsRendering) throw new InvalidOperationException ("Render is in progress");
+		object updatedLock = new ();
+		bool isUpdated = false;
+		RectangleI updatedArea = RectangleI.Zero;
 
-		TaskCompletionSource<CompletionInfo> newCompletionSource = new ();
-		completion_source = newCompletionSource;
+		CancellationTokenSource cts = new ();
 
 		Debug.WriteLine ("AsyncEffectRenderer.Start ()");
 
@@ -129,78 +78,61 @@ internal sealed class AsyncEffectRenderer : IDisposable
 		// So a copy is made for the render.
 		BaseEffect effectClone = effect.Clone ();
 
-		object updated_lock = new ();
-		bool is_updated = false;
-
 		ConcurrentQueue<Exception> renderExceptions = new ();
 
-		ConcurrentQueue<RectangleI> targetTiles = new (
+		ConcurrentQueue<RectangleI> queuedTiles = new (
 			settings.EffectIsTileable
 			? settings.RenderBounds.ToRows () // If effect is tileable, render each row in parallel.
 			: [settings.RenderBounds]); // If the effect isn't tileable, there is a single tile for the entire render bounds
 
-		queued_tiles = targetTiles;
-		tiles_count = targetTiles.Count;
+		int tilesCount = queuedTiles.Count;
 
-		CancellationToken cancellationToken = ReplaceCancellationSource ();
+		Debug.WriteLine ("AsyncEffectRenderer.Start () Render starting.");
 
-		Debug.WriteLine ("AsyncEffectRenderer.Start () Render starting."); // TODO: Show some kind of ID, perhaps the address
+		var tasks =
+			Enumerable.Range (0, settings.ThreadCount)
+			.Select (_ => Task.Run (RenderNextTile));
 
-		// Start slave render threads.
-		var slaves =
-			Enumerable.Range (0, settings.ThreadCount - 1)
-			.Select (_ => StartRenderThread (RenderNextTile))
-			.ToImmutableArray ();
+		var aggregateTask =
+			Task
+			.WhenAll (tasks)
+			.ContinueWith (
+				_ => new CompletionInfo (
+					WasCanceled: cts.Token.IsCancellationRequested,
+					Errors: [.. renderExceptions]
+				)
+			);
 
-		// Start the master render thread.
-		Thread master = StartRenderThread (
-			new ThreadStart (RenderNextTile) // Do part of the rendering on the master thread.
-			+ new ThreadStart (CoordinateSlaveThreads));
-
-		// Start timer used to periodically fire update events on the UI thread.
-		timer_tick_id = GLib.Functions.TimeoutAdd (
-			0,
-			(uint) settings.UpdateMillis,
-			HandleTimerTick);
+		return new (
+			aggregateTask,
+			cts,
+			TryConsumeBounds,
+			GetProgress);
 
 		// ---------------
 		// === Methods ===
 		// ---------------
 
-		Thread StartRenderThread (ThreadStart callback)
+
+		double GetProgress ()
 		{
-			Thread result = new (callback) { Priority = settings.ThreadPriority };
-			result.Start ();
-			return result;
+			if (tilesCount == 0) return 0;
+			int dequeued = tilesCount - queuedTiles.Count;
+			return dequeued / (double) tilesCount;
 		}
 
-		void CoordinateSlaveThreads ()
+		bool TryConsumeBounds (out RectangleI bounds)
 		{
-			// Wait for slave threads to complete.
-			foreach (var slave in slaves)
-				slave.Join ();
-
-			// Change back to the UI thread to notify of completion.
-			GLib.Functions.TimeoutAdd (
-				0,
-				0,
-				() => {
-					HandleTimerTick ();
-
-					if (timer_tick_id > 0)
-						GLib.Source.Remove (timer_tick_id);
-
-					timer_tick_id = 0;
-
-					CompletionInfo completion = new (
-						WasCanceled: cancellationToken.IsCancellationRequested,
-						Errors: [.. renderExceptions]);
-
-					newCompletionSource.SetResult (completion);
-
-					return false; // don't call the timer again
+			lock (updatedLock) {
+				if (!isUpdated) {
+					bounds = default;
+					return false;
 				}
-			);
+
+				bounds = updatedArea;
+				isUpdated = false;
+				return true;
+			}
 		}
 
 		// Runs on a background thread.
@@ -209,89 +141,33 @@ internal sealed class AsyncEffectRenderer : IDisposable
 			// Fetch the next tile index and render it.
 			while (true) {
 
-				if (cancellationToken.IsCancellationRequested) return;
-				if (!queued_tiles.TryDequeue (out RectangleI tileBounds)) return;
+				if (cts.Token.IsCancellationRequested) return;
+				if (!queuedTiles.TryDequeue (out RectangleI tileBounds)) return;
 				try {
 					// NRT - These are set in Start () before getting here
-					if (!cancellationToken.IsCancellationRequested) {
+					if (!cts.Token.IsCancellationRequested) {
 						dest.Flush ();
-						effectClone.Render (source, dest, stackalloc[] { tileBounds });
+						effectClone.Render (source, dest, [tileBounds]);
 						dest.MarkDirty (tileBounds);
 					}
 				} catch (Exception ex) {
 					renderExceptions.Enqueue (ex);
 				}
 
-				// Ignore completions of tiles after a cancel or from a previous render.
-				if (!IsRendering || cancellationToken.IsCancellationRequested)
+				// Ignore completions of tiles after a cancel
+				if (cts.Token.IsCancellationRequested)
 					return;
 
 				// Update bounds to be shown on next expose.
-				lock (updated_lock) {
-					if (is_updated) {
-						updated_area = RectangleI.Union (tileBounds, updated_area);
+				lock (updatedLock) {
+					if (isUpdated) {
+						updatedArea = RectangleI.Union (tileBounds, updatedArea);
 					} else {
-						is_updated = true;
-						updated_area = tileBounds;
+						isUpdated = true;
+						updatedArea = tileBounds;
 					}
 				}
 			}
 		}
-
-		// Called on the UI thread.
-		bool HandleTimerTick ()
-		{
-			Debug.WriteLine (DateTime.Now.ToString ("HH:mm:ss:ffff") + " Timer tick.");
-
-			RectangleI bounds;
-
-			lock (updated_lock) {
-
-				if (!is_updated)
-					return true;
-
-				is_updated = false;
-
-				bounds = updated_area;
-			}
-
-			if (IsRendering && !cancellationToken.IsCancellationRequested)
-				Updated?.Invoke (Progress, bounds);
-
-			return true;
-		}
-	}
-
-	internal Task<CompletionInfo> Finish (bool cancel)
-	{
-		if (cancel) {
-			Debug.WriteLine ("AsyncEffectRenderer.Cancel ()");
-			cancellation_source.Cancel ();
-		}
-		return completion_source.Task;
-	}
-
-	internal delegate void UpdateHandler (
-		double progress,
-		RectangleI updatedBounds);
-
-	public event UpdateHandler? Updated;
-
-	public void Dispose ()
-	{
-		if (timer_tick_id > 0)
-			GLib.Source.Remove (timer_tick_id);
-
-		timer_tick_id = 0;
-	}
-
-	CancellationToken ReplaceCancellationSource ()
-	{
-		CancellationTokenSource newSource = new ();
-		CancellationTokenSource oldSource = cancellation_source;
-		oldSource.Cancel (); // Safe to call multiple times
-		oldSource.Dispose (); // Not safe to call multiple times, so this is the only place it should be called
-		cancellation_source = newSource;
-		return newSource.Token;
 	}
 }