+++ Pinta.Core/Managers/PaletteManager.cs
@@ -36,8 +36,9 @@ public interface IPaletteService
 {
 	Color PrimaryColor { get; set; }
 	Color SecondaryColor { get; set; }
+	Palette CurrentPalette { get; }
+	ReadOnlyCollection<Color> RecentlyUsedColors { get; }
 	void SetColor (bool setPrimary, Color color, bool addToRecent = true);
-
 	public event EventHandler? PrimaryColorChanged;
 	public event EventHandler? SecondaryColorChanged;
 }