+++ Pinta.Effects/CoreEffectsExtension.cs
@@ -60,6 +60,7 @@ internal sealed class CoreEffectsExtension : IExtension
 		PintaCore.Effects.RegisterEffect (new AlignObjectEffect (services));
 		PintaCore.Effects.RegisterEffect (new BulgeEffect (services));
 		PintaCore.Effects.RegisterEffect (new CausticsEffect (services));
+		PintaCore.Effects.RegisterEffect (new CellsEffect (services));
 		PintaCore.Effects.RegisterEffect (new CloudsEffect (services));
 		PintaCore.Effects.RegisterEffect (new DentsEffect (services));
 		PintaCore.Effects.RegisterEffect (new DitheringEffect (services));
@@ -112,6 +113,7 @@ internal sealed class CoreEffectsExtension : IExtension
 		PintaCore.Effects.UnregisterInstanceOfEffect<AddNoiseEffect> ();
 		PintaCore.Effects.UnregisterInstanceOfEffect<AlignObjectEffect> ();
 		PintaCore.Effects.UnregisterInstanceOfEffect<BulgeEffect> ();
+		PintaCore.Effects.UnregisterInstanceOfEffect<CellsEffect> ();
 		PintaCore.Effects.UnregisterInstanceOfEffect<CausticsEffect> ();
 		PintaCore.Effects.UnregisterInstanceOfEffect<CloudsEffect> ();
 		PintaCore.Effects.UnregisterInstanceOfEffect<DentsEffect> ();