+++ Pinta/MainWindow.cs
@@ -26,7 +26,6 @@
 
 using System;
 using System.Linq;
-using System.Threading;
 using Mono.Addins;
 using Pinta.Core;
 using Pinta.Docking;
@@ -98,7 +97,7 @@ internal sealed class MainWindow
 			setupService.RegisterRepositories (true);
 
 		// Look out for any changes in extensions
-		main_thread_id = Thread.CurrentThread.ManagedThreadId;
+		main_thread_id = Environment.CurrentManagedThreadId;
 		AddinManager.AddExtensionNodeHandler (typeof (IExtension), OnExtensionChanged);
 
 		// Load the user's previous settings
@@ -147,10 +146,11 @@ internal sealed class MainWindow
 	{
 		var view = (DocumentViewContent) e.Item;
 
-		if (PintaCore.Workspace.OpenDocuments.IndexOf (view.Document) < 0)
+		int index = PintaCore.Workspace.OpenDocuments.IndexOf (view.Document);
+		if (index < 0)
 			return;
 
-		PintaCore.Actions.Window.SetActiveDocument (view.Document);
+		PintaCore.Workspace.SetActiveDocument (index);
 		PintaCore.Actions.File.Close.Activate ();
 
 		if (PintaCore.Workspace.OpenDocuments.IndexOf (view.Document) < 0)
@@ -169,7 +169,8 @@ internal sealed class MainWindow
 
 		var view = (DocumentViewContent) item;
 
-		PintaCore.Actions.Window.SetActiveDocument (view.Document);
+		int index = PintaCore.Workspace.OpenDocuments.IndexOf (view.Document);
+		PintaCore.Workspace.SetActiveDocument (index);
 		((CanvasWindow) view.Widget).Canvas.Cursor = PintaCore.Tools.CurrentTool?.CurrentCursor;
 	}
 
@@ -208,8 +209,7 @@ internal sealed class MainWindow
 			if (canvasHasBeenShown)
 				return;
 
-			GLib.Functions.TimeoutAdd (
-				0,
+			GLib.Functions.IdleAdd (
 				0,
 				() => {
 					ZoomToWindow_Activated (o, e);
@@ -294,7 +294,7 @@ internal sealed class MainWindow
 	{
 		// Run synchronously if invoked from the main thread, e.g. when loading
 		// addins at startup we require them to be immediately loaded.
-		if (Thread.CurrentThread.ManagedThreadId == main_thread_id)
+		if (Environment.CurrentManagedThreadId == main_thread_id)
 			UpdateExtension (args);
 		else {
 			// Otherwise, schedule the addin to be loaded/unloaded from the main thread
@@ -628,7 +628,7 @@ internal sealed class MainWindow
 
 	private void ZoomToSelection_Activated (object sender, EventArgs e)
 	{
-		PintaCore.Workspace.ActiveWorkspace.ZoomToCanvasRectangle (PintaCore.Workspace.ActiveDocument.Selection.SelectionPath.GetBounds ().ToDouble ());
+		PintaCore.Workspace.ActiveWorkspace.ZoomToCanvasRectangle (PintaCore.Workspace.ActiveDocument.Selection.GetBounds ());
 	}
 
 	private void ZoomToWindow_Activated (object sender, EventArgs e)